+++ LUTCalc/js/colourspaceworker.js
@@ -1,5 +1,5 @@
 var lutCSWorker;
 if (typeof importScripts === 'function') {
-	importScripts('colourspace.js','lut.js','ring.js','brent.js');
+	importScripts('colourspaceworkerscripts.min.js');
 	lutCSWorker = new LUTCSWorker();
 }